cyclotomic: the ring on its coefficient width; tiers 32/64 and 64/128
Real30 becomes Real30T<Coef>, computing in the width of its
coefficients under the same magnitude guard, and the kis metric policy
takes its storage type and its ring separately. Two instantiations are
named: CyclotomicMetric32, 32-bit storage over the 64-bit ring, the
batch tier (every fullerene through about C15,000); and
CyclotomicMetric64, 64-bit storage over the 128-bit ring of
cyclotomic_wide.hh, host only, for single isomers beyond that envelope
(through N of order 10^9), which the library's owner path now runs.
The 64/64 names remain as the parallel port's transition alias.

The batch tier's headers spell no integer wider than 64 bits: the sign
evaluation runs on 32-bit limbs and the modular division folds the two
words of a product with 64-bit arithmetic. The 128-bit ring's oracle
runs at 576 bits (the Liouville bound asks for 525), its constant
derived by the same exact bisection and verified through psi at compile
time. A product poisoning inside the diamond constructor is reported as
Poisoned, the input leaving the envelope, not as an inconsistency.

Verified by the property suite (162 checks, a 128-bit group added),
test_cyclotomic_idt on both tiers, every isomer C20 to C100 (batch tier
against the library build: same canonical word, no refusal), the (5,0)
series through C1020 on both tiers, the C21020 tube refusing by name on
32/64 and computing on 64/128, and the C100000 tube computing on 64/128
with its longest edge exactly (3n-1)/2 = 14996.5.

+++ include/fullerenes/cyclotomic_ambient.hh
@@ -65,10 +65,9 @@ inline constexpr auto kPow60 = make_pow_table60();
 
 // ---------------------------------------------------------------------------
 // The ring Z[zeta_60] (its rank-16 product folds through kPow60 with
-// per-coordinate amplification <= 58 B^2, the guard's constant; the
-// accumulation is 128-bit -- this is host-tier scaffolding).
+// per-coordinate amplification <= 58 B^2, the guard's constant).
 // ---------------------------------------------------------------------------
-struct Zeta60 : CheckedCoeffRing<Zeta60, 16, __int128> {
+struct Zeta60 : CheckedCoeffRing<Zeta60, 16, long long> {
   static constexpr int kAmplification = 58;
   static const std::array<long long, 16>& pow_row(int k) {
     return detail::kPow60[k];
